add `locales/util` export
