+++ packages/app/src/pages/Rewards/Overview/index.tsx
@@ -20,7 +20,7 @@ import { useValidators } from 'contexts/Validators/ValidatorEntries'
 import { useAverageRewardRate } from 'hooks/useAverageRewardRate'
 import { Balance } from 'library/Balance'
 import { CardWrapper } from 'library/Card/Wrappers'
-import { formatFiatCurrency } from 'locales'
+import { formatFiatCurrency } from 'locales/util'
 import { AverageRewardRate } from 'pages/Overview/Stats/AveragelRewardRate'
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
