+++ packages/locales/package.json
@@ -4,6 +4,10 @@
   "license": "GPL-3.0-only",
   "type": "module",
   "main": "src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./util": "./src/util/index.ts"
+  },
   "scripts": {
     "clear": "rm -rf build tsconfig.tsbuildinfo dist",
     "reset": "yarn run clear &&  rm -rf node_modules yarn.lock && yarn",
