+++ packages/locales/src/util/index.ts
@@ -1,0 +1,5 @@
+// Copyright 2025 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+export * from './currency'
+export * from './language'
